% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakademic_theme.R
\name{hakademic_theme}
\alias{hakademic_theme}
\title{A Hakai theme for ggplot2 for hakademics}
\usage{
hakademic_theme(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{Base size}

\item{base_family}{Base family}
}
\description{
A Hakai theme for ggplot2 for hakademics
}
\examples{
p <- ggplot2::ggplot(mtcars) +
  ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = factor(gear))) +
  ggplot2::facet_wrap(~am)
p + hakademic_theme()

}
